[fix] move command to sub directory
